backend/ml_models/requirements.txt
ultralytics==8.0.196
opencv-python==4.8.1.78
requests==2.31.0
Pillow==10.0.1
numpy==1.24.3
python-dotenv==1.0.0
groq==0.4.2
torch==2.1.0
torchvision==0.16.0
streamlit==1.28.1 